requirements.txt
PyHamcrest==1.8.0
PyPubSub==3.3.0
astroid==1.0.1
coverage==3.7.1
logilab-common==0.60.1
mock==1.0.1
nose==1.3.0
pep8==1.4.6
pylint==1.1.0
wsgiref==0.1.2
